core: compilation fixes

+++ core/kos_module.c
@@ -441,7 +441,7 @@ static int predefine_globals(KOS_CONTEXT    ctx,
 
         TRY(kos_compiler_predefine_global(program,
                                           cpath.buffer,
-                                          cpath.size - 1,
+                                          (uint16_t)cpath.size - 1,
                                           (int)GET_SMALL_INT(KOS_get_walk_value(walk.o)),
                                           is_repl ? 0 : 1));
     }
@@ -455,7 +455,7 @@ static int predefine_globals(KOS_CONTEXT    ctx,
 
         TRY(KOS_string_to_cstr_vec(ctx, KOS_get_walk_key(walk.o), &cpath));
 
-        TRY(kos_compiler_predefine_module(program, cpath.buffer, cpath.size - 1,
+        TRY(kos_compiler_predefine_module(program, cpath.buffer, (uint16_t)cpath.size - 1,
                                           (int)GET_SMALL_INT(KOS_get_walk_value(walk.o))));
     }
 
@@ -881,7 +881,7 @@ int kos_comp_walk_globals(void                          *vframe,
         TRY(KOS_string_to_cstr_vec(ctx, KOS_get_walk_key(walk.o), &name));
 
         TRY(callback(name.buffer,
-                     (unsigned)name.size - 1U,
+                     (uint16_t)name.size - 1U,
                      module_idx,
                      (int)GET_SMALL_INT(KOS_get_walk_value(walk.o)),
                      cookie));
@@ -1113,7 +1113,7 @@ static int print_func(void *cookie, uint32_t func_index)
 
 static int compile_module(KOS_CONTEXT ctx,
                           KOS_OBJ_ID  module_obj,
-                          int         module_idx,
+                          uint16_t    module_idx,
                           const char *data,
                           unsigned    data_size,
                           int         is_repl)
@@ -1142,7 +1142,7 @@ static int compile_module(KOS_CONTEXT ctx,
         program.num_constants = (int)KOS_get_array_size(OBJPTR(MODULE, module.o)->constants);
     kos_parser_init(&parser,
                     &program.allocator,
-                    (unsigned)module_idx,
+                    module_idx,
                     data,
                     data + data_size);
 
@@ -1637,13 +1637,13 @@ static KOS_OBJ_ID import_and_run(KOS_CONTEXT ctx,
     }
 
     /* Compile module source to bytecode */
-    TRY(compile_module(ctx, module.o, module_idx, data, data_size, 0));
+    TRY(compile_module(ctx, module.o, (uint16_t)module_idx, data, data_size, 0));
 
     /* Free file buffer */
     kos_unload_file(&file_buf);
 
     /* Put module on the list */
-    TRY(KOS_array_write(ctx, inst->modules.modules, module_idx, module.o));
+    TRY(KOS_array_write(ctx, inst->modules.modules, (uint16_t)module_idx, module.o));
     TRY(KOS_set_property(ctx, inst->modules.module_names, actual_module_name.o, TO_SMALL_INT(module_idx)));
 
     /* Run module */
